+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## [Unreleased]
+
+### Added
+- Support for importing WPL playlist files and play them within Files
+  [#1153](https://github.com/owncloud/music/issues/1153)
+  * MIME mapping for the file is needed which can be added with `occ music:register-mime-types`
+
+### Changed
+- In the Files app, use the Music app icon on the file actions menu item "Play"
+
+### Fixed
+
 ## 2.1.4 - 2025-05-10
 
 ### Fixed